DSInteger now takes a long in the constructor as opposed to an int

+++ info/paulmchugh/stellariseditor/datatypes/DSInteger.java
@@ -9,7 +9,7 @@ public class DSInteger implements SaveElement
 		this(0);
 	}
 	
-	public DSInteger(int value)
+	public DSInteger(long value)
 	{
 		this.value = value;
 	}
